chore: initialize repository for buptLab-os
This commit sets up the repository for the "Operating Systems" course experiments at Beijing University of Posts and Telecommunications for the 2024-2025 autumn semester taught by Professor Ye Wen.

The repository contains the related code and reports for six course experiments:

1. Process creation and management.
2. Thread creation, management, and communication.
3. Inter-process communication.
4. Synchronization and mutual exclusion of processes/threads.
5. Multi-core multi-threaded programming and performance analysis.
6. Observation of process address space.

Experiments are primarily conducted on Ubuntu 24.04, with necessary configuration steps detailed in the experiment reports.

The following files are added as part of this commit:

- .editorconfig
- .github/ISSUE_TEMPLATE/bug_report.md
- .github/ISSUE_TEMPLATE/custom.md
- .github/ISSUE_TEMPLATE/feature_request.md
- .github/pull_request_template.md
- .github/workflows/commit-check.yml
- CODE_OF_CONDUCT.md
- CONTRIBUTING.md
- LICENSE
- README.md
- SECURITY.md